combine entity and trigger map maker lists
remove placeable trigger class -- just use MapMakerEntityMatcher
