+++ rec/maps/Puppy/Puppy.txt
@@ -1,0 +1,33 @@
+{
+  "areas": [],
+  "mapTransitions": [],
+  "NPCs": [],
+  "items": [
+    {
+      "item": "Water Stone",
+      "location": {
+        "x": 4,
+        "y": 7
+      },
+      "triggerName": "Puppy_Item_WaterStone_01"
+    }
+  ],
+  "miscEntities": [],
+  "events": [],
+  "wildBattles": [
+    {
+      "name": "nanana",
+      "encounterRate": "VERY_COMMON",
+      "pokemon": [
+        "Mew 1-100 100%"
+      ],
+      "location": [
+        {
+          "x": 9,
+          "y": 7
+        }
+      ],
+      "triggerName": "Puppy_WildBattle_nanana_01"
+    }
+  ]
+}
