+++ src/pattern/EventMatcher.java
@@ -1,10 +1,7 @@
 package pattern;
 
 import map.entity.EntityAction;
-import map.entity.EntityAction.TriggerAction;
-import map.triggers.TriggerType;
 import mapMaker.model.TriggerModel.TriggerModelType;
-import util.JsonUtils;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -42,23 +39,5 @@ public class EventMatcher extends MultiPointEntityMatcher {
         }
 
         return entityActions;
-    }
-
-    public boolean isWildBattleTrigger() {
-        return this.getWildBattleTriggerContents() != null;
-    }
-
-    public WildBattleMatcher getWildBattleTriggerContents() {
-        for (EntityAction action : this.getActions()) {
-            if (action instanceof TriggerAction) {
-                TriggerAction triggerAction = (TriggerAction)action;
-                if (triggerAction.getTriggerType() == TriggerType.WILD_BATTLE) {
-                    String contents = triggerAction.getTriggerContents(this.name);
-                    return JsonUtils.deserialize(contents, WildBattleMatcher.class);
-                }
-            }
-        }
-
-        return null;
     }
 }
